==> Code/1a-process_expression_data_ECOLI.ipynb <==
%matplotlib inline

import pandas as pd
import numpy as np
from scipy import stats

import statsmodels.formula.api as smf

from matplotlib import pyplot as plt

# Import data

This data comes from this paper: https://www.nature.com/articles/s41467-019-13483-w

It's a big excel file with expression and growth rate data in different sheets and this code will use a few of these sheets.

df = pd.ExcelFile(r'../Data/raw_data/41467_2019_13483_MOESM4_ESM.xlsx')

## Basic metadata clean up and data subsetting

After some manual inspection, looking at distributions, etc. we decided to exclude samples where:

1. The growth rate data was unknown


2. The growth rate data was reported as zero. This one might seem strange but it's is a little unclear in general if that's possible/true to have zero growth rate. It is possible that these were stationary phase cultures but equally likely from my stand-point that these are errors in the table.


3. Really poor alignment (perhaps indicating some overall contamination)

meta_df = df.parse('Metadata') ###This grabs the sheet that contains information about the samples
print(meta_df.shape)
meta_df = meta_df[meta_df['Growth Rate (1/hr)'].isnull() == False]
print(meta_df.shape)
meta_df = meta_df[meta_df['Growth Rate (1/hr)'] > 0.0]
print(meta_df.shape)
meta_df = meta_df[meta_df['Alignment'] > 80]
print(meta_df.shape)
meta_df.head()

## Read in the expression data

exp_df = df.parse('Expression Data', index_col='log-TPM')
print(exp_df.shape)
exp_df = exp_df[meta_df['Sample ID']] ###Only grab the columns corresponding to the samples identified above
print(exp_df.shape)
assert list(meta_df['Sample ID']) == list(exp_df.columns) ###Check our work
exp_df.head()

**Manual inspection found some weird and highly implausible/impossible duplicate column/s, so we'll make a note of that here and deal with it later**

exp_df[[col for col in exp_df.columns if 'pal__lyx_ale' in col]].head()

# Averaging gene expression values across replicates to clean up/simplify the data

## First getting gene expression averages between replicates and creating a new `dataframe` to hold this information

Thus just exploits the fact that replicates are denoted by "__x" in the sample names so we identify these and average them when possible. 

unique_cols = set(exp_df.columns.str[:-3])
new_exp_df = pd.DataFrame()

for i in unique_cols:
    new_exp_df[i] = exp_df[[col for col in exp_df.columns if col[:-3]==i]].mean(axis=1)
print(new_exp_df.shape)
new_exp_df.head()

**Double checking the work**

Just making sure somethings add up here by taking an example column/condition and looking at the replicate values

example_col = list(unique_cols)[1]

exp_df[[col for col in exp_df.columns if col[:-3]==example_col]].head()

And their average

new_exp_df[[example_col]].head()

## Dealing with the weird duplicate column/s

Time to kill any completely identical columns, these must be bugs on the data end and even though I could in theory keep one, their growth rate value is unclear (since it differs)

###Get an all-to-all correlation matrix between gene expression values
temp_corr = new_exp_df.corr(method='spearman')

###And make this into a symmetric dataframe
temp_df = pd.DataFrame(
    np.where(np.equal(*np.indices(temp_corr.shape)), np.nan, temp_corr.values),
    temp_corr.index, temp_corr.columns
)
print(temp_df.shape)
temp_df.head()

**Identify columns that contain a value of "1." since this indicates they have a perfect correlation**

temp_df[temp_df.values==1]

Get rid of them both since there is obviously an error here somewhere

new_exp_df.drop(['pal__lyx_ale2', 'pal__lyx_ale4'], axis=1, inplace=True)
print(new_exp_df.shape)

# Average the growth rates across these replicates in the metadata as well

###Assign a unique id that removes the replicate information
meta_df['Simple_sample_id'] = meta_df['Sample ID'].str[:-3]
print(meta_df.shape)
###Group according to this new id
group_cols = ['Simple_sample_id']
###For these numeric columns I'll take the mean
metric_cols_a = ['Temperature (C)', 'pH', 'Growth Rate (1/hr)', 'Alignment']
aggs_a = meta_df.groupby(group_cols)[metric_cols_a].mean()
###And for these I'll just grab the count
metric_cols_b = ['Replicate #', 'Biological Replicates']
aggs_b = meta_df.groupby(group_cols)['Replicate #'].count()

###Drop the columns from the original dataframe (we'll add them back in later)
meta_df.drop(metric_cols_a, axis=1, inplace=True)
meta_df.drop(metric_cols_b, axis=1, inplace=True)
###And duplicates
meta_df.drop_duplicates(subset=group_cols, keep='first', inplace=True)

###Now merge the main dataframe with the grouped ones
meta_df = meta_df.merge(right=aggs_a, right_index=True, left_on=group_cols, how='right')
print(meta_df.shape)
meta_df = meta_df.merge(right=aggs_b, right_index=True, left_on=group_cols, how='right')
print(meta_df.shape)
meta_df.head()

**And get rid of those problematic samples from this dataframe as well**

print(meta_df.shape)
meta_df = meta_df[meta_df['Sample ID'].str.contains('pal__lyx_ale2__')==False]
print(meta_df.shape)
meta_df = meta_df[meta_df['Sample ID'].str.contains('pal__lyx_ale4__')==False]
print(meta_df.shape)

**Sum should equal the original shape!**

meta_df['Replicate #'].value_counts()

**Make sure that the columns line up when matching across these two dataframes** 

new_exp_df = new_exp_df[meta_df['Simple_sample_id']]

assert all(new_exp_df.columns == meta_df['Simple_sample_id'])

**And add the doubling time just for good measure**

Which is just a slight transformation of growth rate

meta_df['Doubling_time'] = np.log(2)/meta_df['Growth Rate (1/hr)']

# Construct a third `dataframe` containing gene expression data summary stats

Strictly speaking this isn't super necessary but might as well do it now to get it done and over with

exp_summary_df = new_exp_df.apply(pd.DataFrame.describe, axis=1)

**The % signs seem to cause some problems down the road so lets remove them**

col_listy = []
for col in exp_summary_df.columns:
    if '%' not in col:
        col_listy.append(col)
    else:
        col_listy.append(col.replace('%', '_percentile'))
print(col_listy)
exp_summary_df.columns = col_listy

**And add some other dispersion metrics**

exp_summary_df['cv'] = exp_summary_df['std']/exp_summary_df['mean']
exp_summary_df['noise'] = exp_summary_df['std'].pow(2)/exp_summary_df['mean'].pow(2)

**Finally, adding some of (what we think are) the cool new variables to consider**

slopes = []
pearsons = []
spearmans = []
for gene in exp_summary_df.index:
    a, b, c, d, e = stats.linregress(new_exp_df.loc[gene], meta_df['Growth Rate (1/hr)'])
    slopes.append(a)
    pearsons.append(c)
    rho, p = stats.spearmanr(new_exp_df.loc[gene], meta_df['Growth Rate (1/hr)'])
    spearmans.append(rho)
    
exp_summary_df['lin_slope'] = slopes
exp_summary_df['lin_r'] = pearsons
exp_summary_df['spearmans_rho'] = spearmans

# Save some files

This was the whole point of all the code above. Should have some straightforward data tables now

print(new_exp_df.shape)
print(meta_df.shape)
print(exp_summary_df.shape)

new_exp_df.to_csv('../Data/processed_data/processed_expression_ecoli.tsv', sep='\t')
meta_df.to_csv('../Data/processed_data/processed_metadata_ecoli.tsv', sep='\t')
exp_summary_df.to_csv('../Data/processed_data/processed_expression_summary_ecoli.tsv', sep='\t')

# Get a thinned down dataset as a robustness check

Since conditions are correlated with one another, I'm constructing a more sparsely populated dataset where inter-condition correlations are minimized using a greedy algorithm. 

The algorithm works as follows:
1. Find which two conditions are the most highly correlated across the entire all-to-all correlation matrix
2. Randomly delete one of the two conditions in question
3. Iterate to (1)
4. Stop after reaching a pre-defined final dataset size

import random
random.seed(42)

cond_corr_mat = new_exp_df.corr(method='spearman')
cond_corr_mat.head()

**The `final_data_size` is of course completely arbitrary. But the point is to get rid of some correlated data so it does the trick.**

final_data_size = 30
n_to_prune = cond_corr_mat.shape[1] - final_data_size

np.fill_diagonal(cond_corr_mat.values, np.nan)
to_prune = []
for i in range(n_to_prune):
    tempy = cond_corr_mat.loc[[i for i in cond_corr_mat.columns if i not in to_prune]][[i for i in cond_corr_mat.columns if i not in to_prune]].max()
    to_prune.append(random.choice([tempy.sort_values(ascending=False).index[0],\
                                   tempy.sort_values(ascending=False).index[1]]))
np.fill_diagonal(cond_corr_mat.values, 1.)

**Remove relevant rows from two of the dataframes**

to_keep = [i for i in cond_corr_mat.columns if i not in to_prune]
new_exp_sprs_df = new_exp_df[to_keep]
meta_sprs_df = meta_df[meta_df['Simple_sample_id'].isin(to_keep)]

**And regenerate the summary dataframe to only use this thinned-down set of samples**

exp_summary_sprs_df = new_exp_sprs_df.apply(pd.DataFrame.describe, axis=1)
col_listy = []
for col in exp_summary_sprs_df.columns:
    if '%' not in col:
        col_listy.append(col)
    else:
        col_listy.append(col.replace('%', '_percentile'))
print(col_listy)
exp_summary_sprs_df.columns = col_listy

exp_summary_sprs_df['cv'] = exp_summary_sprs_df['std']/exp_summary_sprs_df['mean']
exp_summary_sprs_df['noise'] = exp_summary_sprs_df['std'].pow(2)/exp_summary_sprs_df['mean'].pow(2)

slopes = []
pearsons = []
spearmans = []
for gene in exp_summary_sprs_df.index:
    a, b, c, d, e = stats.linregress(new_exp_sprs_df.loc[gene], meta_sprs_df['Growth Rate (1/hr)'])
    slopes.append(a)
    pearsons.append(c)
    rho, p = stats.spearmanr(new_exp_sprs_df.loc[gene], meta_sprs_df['Growth Rate (1/hr)'])
    spearmans.append(rho)
    
exp_summary_sprs_df['lin_slope'] = slopes
exp_summary_sprs_df['lin_r'] = pearsons
exp_summary_sprs_df['spearmans_rho'] = spearmans

print(new_exp_sprs_df.shape)
print(meta_sprs_df.shape)
print(exp_summary_sprs_df.shape)

new_exp_sprs_df.to_csv('../Data/processed_data/processed_expression_ecoli_SPARSE.tsv', sep='\t')
meta_sprs_df.to_csv('../Data/processed_data/processed_metadata_ecoli_SPARSE.tsv', sep='\t')
exp_summary_sprs_df.to_csv('../Data/processed_data/processed_expression_summary_ecoli_SPARSE.tsv', sep='\t')

## Create neutral data sets for additional robustness check
**Limit analysis to non-ALE (adaptive laboratory environment) strains to control for selection**
**First, remove ALE strains in data sets**

meta_ntrl_df = meta_df[meta_df['Evolved Sample'].str.contains('No')==True]
print(meta_ntrl_df.shape)

new_exp_ntrl_df = new_exp_df[meta_ntrl_df['Simple_sample_id']]
print(new_exp_ntrl_df.shape)

**And regenerate the summary dataframe to only use this non-ALE set of samples**

exp_summary_ntrl_df = new_exp_ntrl_df.apply(pd.DataFrame.describe, axis=1)
col_listy = []
for col in exp_summary_ntrl_df.columns:
    if '%' not in col:
        col_listy.append(col)
    else:
        col_listy.append(col.replace('%', '_percentile'))
print(col_listy)
exp_summary_ntrl_df.columns = col_listy

exp_summary_ntrl_df['cv'] = exp_summary_ntrl_df['std']/exp_summary_ntrl_df['mean']
exp_summary_ntrl_df['noise'] = exp_summary_ntrl_df['std'].pow(2)/exp_summary_ntrl_df['mean'].pow(2)

slopes = []
pearsons = []
spearmans = []
for gene in exp_summary_ntrl_df.index:
    a, b, c, d, e = stats.linregress(new_exp_ntrl_df.loc[gene], meta_ntrl_df['Growth Rate (1/hr)'])
    slopes.append(a)
    pearsons.append(c)
    rho, p = stats.spearmanr(new_exp_ntrl_df.loc[gene], meta_ntrl_df['Growth Rate (1/hr)'])
    spearmans.append(rho)
    
exp_summary_ntrl_df['lin_slope'] = slopes
exp_summary_ntrl_df['lin_r'] = pearsons
exp_summary_ntrl_df['spearmans_rho'] = spearmans

print(new_exp_ntrl_df.shape)
print(meta_ntrl_df.shape)
print(exp_summary_ntrl_df.shape)

**Save the files for use later in the pipeline**

new_exp_ntrl_df.to_csv('../Data/processed_data/processed_expression_ecoli_NEUTRAL.tsv', sep='\t')
meta_ntrl_df.to_csv('../Data/processed_data/processed_metadata_ecoli_NEUTRAL.tsv', sep='\t')
exp_summary_ntrl_df.to_csv('../Data/processed_data/processed_expression_summary_ecoli_NEUTRAL.tsv', sep='\t')

**Limit analysis to non-ALE, non-mutants, and non-knock-outs**

**First, filter out mutants and knock-outs from neutral data sets**

# remove knock-outs
temp_meta_ntrl_wt_df = meta_ntrl_df[meta_ntrl_df['Strain Description'].str.contains('del')==False]
print(temp_meta_ntrl_wt_df.shape)

# remove mutants
non_mut = ['Escherichia coli K-12 MG1655', 'Escherichia coli BW25113']
meta_ntrl_wt_df = temp_meta_ntrl_wt_df[temp_meta_ntrl_wt_df['Strain Description'].isin(non_mut)]
print(meta_ntrl_wt_df.shape)

# remove knock-outs and mutants
new_exp_ntrl_wt_df = new_exp_df[meta_ntrl_wt_df['Simple_sample_id']]
print(new_exp_ntrl_wt_df.shape)

**And regenerate summary stats for WT neutral data set**

exp_summary_ntrl_wt_df = new_exp_ntrl_wt_df.apply(pd.DataFrame.describe, axis=1)
col_listy = []
for col in exp_summary_ntrl_wt_df.columns:
    if '%' not in col:
        col_listy.append(col)
    else:
        col_listy.append(col.replace('%', '_percentile'))
print(col_listy)
exp_summary_ntrl_wt_df.columns = col_listy

exp_summary_ntrl_wt_df['cv'] = exp_summary_ntrl_wt_df['std']/exp_summary_ntrl_wt_df['mean']
exp_summary_ntrl_wt_df['noise'] = exp_summary_ntrl_wt_df['std'].pow(2)/exp_summary_ntrl_wt_df['mean'].pow(2)

slopes = []
pearsons = []
spearmans = []
for gene in exp_summary_ntrl_wt_df.index:
    a, b, c, d, e = stats.linregress(new_exp_ntrl_wt_df.loc[gene], meta_ntrl_wt_df['Growth Rate (1/hr)'])
    slopes.append(a)
    pearsons.append(c)
    rho, p = stats.spearmanr(new_exp_ntrl_wt_df.loc[gene], meta_ntrl_wt_df['Growth Rate (1/hr)'])
    spearmans.append(rho)
    
exp_summary_ntrl_wt_df['lin_slope'] = slopes
exp_summary_ntrl_wt_df['lin_r'] = pearsons
exp_summary_ntrl_wt_df['spearmans_rho'] = spearmans

print(new_exp_ntrl_wt_df.shape)
print(meta_ntrl_wt_df.shape)
print(exp_summary_ntrl_wt_df.shape)

**Save files for use later in pipeline**

new_exp_ntrl_wt_df.to_csv('../Data/processed_data/processed_expression_ecoli_NEUTRAL_WT.tsv', sep='\t')
meta_ntrl_wt_df.to_csv('../Data/processed_data/processed_metadata_ecoli_NEUTRAL_WT.tsv', sep='\t')
exp_summary_ntrl_wt_df.to_csv('../Data/processed_data/processed_expression_summary_ecoli_NEUTRAL_WT.tsv', sep='\t')
