==> Code/1b-process_expression_data_SCER.ipynb <==
%matplotlib inline

import pandas as pd
import numpy as np
from scipy import stats

import statsmodels.formula.api as smf

from matplotlib import pyplot as plt

**Two different sample types are contained in this dataset so we'll try and re-run the notebook to use both of them separately**

Basically, make sure one is commented, run all the code. Then switch the commented line and re-run.

data_to_analyze = 'RNA'
# data_to_analyze = 'PROTEIN'

# Import data

This data comes from here: https://elifesciences.org/articles/65722/figures

df = pd.ExcelFile(r'../Data/raw_data/elife-65722-supp1-v2.xlsx')

## Basic metadata clean up and data subsetting

meta_df = df.parse('Supplementary File 1a', header=1)
print(meta_df.shape)
meta_df.head()

## And the same for the expression data

exp_df = df.parse('Supplementary File 1b', header=1, index_col=1)
print(exp_df.shape)
exp_df.head()

**Select the columns of interest, normalize so that values are ~molecules per million, and then log-scale**

The only tricky thing here is dealing with zero-values (these only exist in the protein dataset). I thought of a number of solutions, ideally we'd drop them but for a lot of things this becomes a huge pain. Instead, we can just set all 0 values to the smallest observed value anywhere in the dataset and poking around this didn't seem to effect things very much.

million=1e6

if data_to_analyze == 'PROTEIN':
    exp_df = exp_df[['{}.{}'.format('prot', i) for i in meta_df['Sample']]]
elif data_to_analyze == 'RNA':
    exp_df = exp_df[['{}.{}'.format('RNA', i) for i in meta_df['Sample']]]

#####
#####Dealing with zero values here, if necessary
current_minimum = exp_df.min().min()
print('Smallest initial value found:', current_minimum)
if current_minimum == 0.:
    print('Making minor adjustments to minimum values to play nice with zeros')
    nonzero_minimum = exp_df[exp_df>0.].min().min()
    exp_df.replace(0., nonzero_minimum, inplace=True)
    print('New minimum value is {}'.format(nonzero_minimum))
else:
    print('No replacement of / dealing with zeros necessary')
print(exp_df.shape)
#####
#####

exp_df.columns = [i for i in meta_df['Sample']]
assert list(meta_df['Sample']) == list(exp_df.columns)
for i in exp_df.columns:
    summy = exp_df[i].sum() ###Get the sum
    exp_df[i] = np.log((exp_df[i]/summy)*million)
    
print(exp_df.shape)
exp_df.head()

###(slowly/inefficiently) checking for exact duplicate columns
for i, col1 in enumerate(exp_df.columns):
    for col2 in exp_df.columns[i+1:]:
        assert list(exp_df[col1])!= list(exp_df[col2])

# Averaging between replicates to clean up the data

## First getting gene expression averages between replicates and creating a new `dataframe`

replicate_groups = {}
for sample in meta_df.index:
    ###Make unique identifiers for the samples
    tempy = '{}_{}_{}_{}'.format(*meta_df.loc[sample][['Nitrogen source',\
                                                     'dilution rate (/h)',\
                                                     'glucose.H2O concentration (g/L)',\
                                                     'nitrogen source concentration (g/L)']])
    try:
        replicate_groups[tempy].append(meta_df.loc[sample]['Sample'])
    except KeyError:
        replicate_groups[tempy] = [meta_df.loc[sample]['Sample']]

for key, val in replicate_groups.items():
    print(key, val)

new_exp_df = pd.DataFrame()

for key, val in replicate_groups.items():
    new_exp_df[key] = exp_df[list(val)].mean(axis=1)
print(new_exp_df.shape)
new_exp_df.head()

## Average the growth rates across these replicates in the metadata as well

for key, vals in replicate_groups.items():
    for numeric_id in vals:
        temp_df = meta_df[meta_df['Sample']==numeric_id]
        assert temp_df.shape[0]==1
        meta_df.at[temp_df.index[0], 'Simple_sample_id'] = key
print(meta_df.shape)
meta_df.head()

print(meta_df.shape)
group_cols = ['Simple_sample_id']
metric_cols_a = ['total RNA content (g/gDW)', 'total protein content (g/gDW)']
aggs_a = meta_df.groupby(group_cols)[metric_cols_a].mean()

meta_df.drop(metric_cols_a, axis=1, inplace=True)
meta_df.drop_duplicates(subset=group_cols, keep='first', inplace=True)
meta_df = meta_df.merge(right=aggs_a, right_index=True, left_on=group_cols, how='right')
print(meta_df.shape)
meta_df.head()

meta_df['Doubling_time'] = np.log(2)/meta_df['dilution rate (/h)']
meta_df['Growth Rate (1/hr)'] = meta_df['dilution rate (/h)']
assert list(meta_df['Simple_sample_id']) == list(new_exp_df.columns)

# Construct a third `dataframe` containing expression data summary stats

exp_summary_df = new_exp_df.apply(pd.DataFrame.describe, axis=1)

col_listy = []
for col in exp_summary_df.columns:
    if '%' not in col:
        col_listy.append(col)
    else:
        col_listy.append(col.replace('%', '_percentile'))
print(col_listy)
exp_summary_df.columns = col_listy

exp_summary_df['cv'] = exp_summary_df['std']/exp_summary_df['mean']
exp_summary_df['noise'] = exp_summary_df['std'].pow(2)/exp_summary_df['mean'].pow(2)

slopes = []
pearsons = []
spearmans = []
for gene in exp_summary_df.index:
    a, b, c, d, e = stats.linregress(new_exp_df.loc[gene], meta_df['dilution rate (/h)']) 
    slopes.append(a)
    pearsons.append(c)
    rho, p = stats.spearmanr(new_exp_df.loc[gene], meta_df['dilution rate (/h)'], nan_policy='omit')
    spearmans.append(rho)
    
exp_summary_df['lin_slope'] = slopes
exp_summary_df['lin_r'] = pearsons
exp_summary_df['spearmans_rho'] = spearmans

# Save some files

print(new_exp_df.shape)
print(meta_df.shape)
print(exp_summary_df.shape)
new_exp_df.to_csv('../Data/processed_data/processed_expression_scer_{}.tsv'.format(data_to_analyze), sep='\t')
meta_df.to_csv('../Data/processed_data/processed_metadata_scer_{}.tsv'.format(data_to_analyze), sep='\t')
exp_summary_df.to_csv('../Data/processed_data/processed_expression_summary_scer_{}.tsv'.format(data_to_analyze), sep='\t')

# Get a thinned down set as a robustness check?

Too few samples to do this step here.
