==> Code/2-compile_gene_metrics.ipynb <==
%matplotlib inline

# Imports

###Standard imports
import pandas as pd
from collections import Counter
from scipy import stats
import numpy as np
import re

###Slightly specialized
from Bio import SeqIO
from Bio.Data import CodonTable

# Define constants for genome extraction

# organism = 'ecoli'
organism = 'scer'

if organism == 'ecoli':
    genome = list(SeqIO.parse('../Data/ecoli_info/NC_000913.3.gb', 'genbank'))
    assert len(genome) == 1
    genome = genome[0]
    
elif organism == 'scer':
    features = list(SeqIO.parse('../Data/scer_info/GCA_000269885.1_ASM26988v1_cds_from_genomic.fna', 'fasta'))
    assert len(features) > 0
    print(len(features))

else:
    raise Exception('Incorrect organism identifier')

# Establish a general dataframe describing various coding sequence fesatures from the genome

loci = []
gene_ids = []
cds_seqs = []
starts = []
stops = []
strands = []

if organism == 'ecoli':
    for feature in genome.features:
        nt_seq_len = len(str(genome.seq))
        if feature.type == 'CDS': ###Caring only about protein coding sequences aka CDS
            try:
                locus_id = feature.qualifiers['locus_tag'][0]
            except:
                locus_id = ''
            try:
                gene_id = feature.qualifiers['gene'][0]
            except:
                gene_id = ''
            start = feature.location.start
            stop = feature.location.end
            strand = feature.location.strand
            assert stop > start
            if strand == 1:
                strand = '+'
            elif strand == -1:
                strand = '-'
            else:
                print('ERROR in strand')
                break
            if strand == '-':
                cds_seq = str(genome.seq[start:stop].reverse_complement())
            elif strand == '+':
                cds_seq = str(genome.seq[start:stop])
            ###Add everyone to growing lists because it's a faster way to set up dataframes
            loci.append(locus_id)
            gene_ids.append(gene_id)
            starts.append(start)
            stops.append(stop)
            strands.append(strand)
            cds_seqs.append(cds_seq)

elif organism == 'scer':
    for feature in features:
        desc_list = re.findall("\[(.*?)\]", feature.description)
        desc_dict = {desc.split('=')[0]:desc.split('=')[1] for desc in desc_list}
        if desc_dict['gbkey'] != 'CDS':
            raise Exception('Somehow found a non-cds entry')
        locus_id = desc_dict['locus_tag']
        gene_id = desc_dict['protein']
        if ('join' in desc_dict['location']) or ('>' in desc_dict['location']) or ('<' in desc_dict['location']):
            start = None
            stop = None
            strand = None
        elif 'complement' in desc_dict['location']:
            location = desc_dict['location'][desc_dict['location'].index('(')+1:].split('..')
            assert len(location)==2
            start = int(location[0])
            stop = int(location[1].rstrip(')'))
            strand = '-'
            assert stop > start

        else:
            location = desc_dict['location'].split('..')
            assert len(location) == 2
            start = int(location[0])
            stop = int(location[1])
            strand = '+'
            assert stop > start

        cds_seq = str(feature.seq)
        
        loci.append(locus_id)
        gene_ids.append(gene_id)
        starts.append(start)
        stops.append(stop)
        strands.append(strand)
        cds_seqs.append(cds_seq)
else:
    raise Exception('Incorrect organism identifier')

df = pd.DataFrame(zip(loci, gene_ids, starts, stops, strands, cds_seqs)) 
df.columns = ['locus_tag', 'gene', 'start_loc', 'stop_loc', 'strand', 'cds_seq']
print(df.shape)
df.head()

if organism == 'ecoli':
    df['cds_len'] = df['stop_loc'] - df['start_loc']
    assert all(df['cds_len'] == df['cds_seq'].str.len())
    print('Max and min cds lengths:', df['cds_len'].max(), '\t', df['cds_len'].min())
elif organism == 'scer':
    df['cds_len'] = df['cds_seq'].str.len()
    print('Max and min cds lengths:', df['cds_len'].max(), '\t', df['cds_len'].min())
else:
    raise Exception('Incorrect organism identifier')

**Keep track of genes that will and won't be useful for downstream analyses**

The only addition at the end should be a `well_behaved` column

print(df.shape)
###First ensure that locus tags are unique and drop duplicates entirely
df = df.drop_duplicates('locus_tag')

###Next set up a simple column that tracks the suitability of each gene
df['well_behaved'] = True

###Ensure that the coding sequence is a multiple of 3 (otherwise, what is happening?)
df.at[df[df['cds_len']%3 != 0].index, 'well_behaved'] = False

###Count the number of canonical bases
for letter in ['A', 'T', 'G', 'C']:
    df[letter] = df['cds_seq'].str.count(letter)
df['ATGC'] = df[['A', 'T', 'G', 'C']].sum(axis=1)
###And ensure that this number equals the coding sequence length
df.at[df[df['cds_len'] != df['ATGC']].index, 'well_behaved'] = False

###Remove those temporary columns
df = df.drop(columns=['A', 'T', 'G', 'C', 'ATGC'])
print(df.shape)

###Remove suspiciously(?) long genes
len_to_cut = df['cds_len'].median()*5
df.at[df[df['cds_len']>len_to_cut].index, 'well_behaved'] = False

print()
print('Well behaved?')
print(df['well_behaved'].value_counts())

df[df['cds_len'] < df['cds_len'].median()/10]

**Write a set of well-behaved coding sequences to run through ROC-SEMPPR (associated R code)**

if organism == 'ecoli':
    with open('../Data/ecoli_info/NC_000913.3.CLEAN.CDS.fasta', 'w') as outfile:
        for index in df[df['well_behaved']==True].index:
            outfile.write('>{}\n'.format(df.loc[index]['locus_tag']))
            outfile.write('{}\n'.format(df.loc[index]['cds_seq']))
elif organism == 'scer':
    with open('../Data/scer_info/SCER.CLEAN.CDS.fasta', 'w') as outfile:
        for index in df[df['well_behaved']==True].index:
            outfile.write('>{}\n'.format(df.loc[index]['locus_tag']))
            outfile.write('{}\n'.format(df.loc[index]['cds_seq']))
else:
    raise Exception('Incorrect organism identifier')

# Add in GC percent data

###Count the number of canonical bases
for letter in ['A', 'T', 'G', 'C']:
    df[letter] = df['cds_seq'].str.count(letter)
df['GC_percent_cds'] = df[['G', 'C']].sum(axis=1)/df[['A', 'T', 'G', 'C']].sum(axis=1)
df = df.drop(columns=['A', 'T', 'G', 'C'])
print(df.shape)

# Add the `ROC-SEMPPR` data 

(After running the `R` code offline)

`ROC-SEMPPR` is defined here:
https://academic.oup.com/gbe/article/7/6/1559/2465799

It is a pretty novel, and in my opinion under-cited / under-used, metric of codon usage bias that is rooted in population genetic theory. The calculations take some time to run, but require only the **set** of all coding sequences in question from a particular genome, with no other *a priori* information necessary. In this sense it's slightly more complex than `iCUB` but uses far less information than many other models that purport to predict gene expression from sequence information.

if organism == 'ecoli':
    rocsemppr_df = pd.read_csv('../Data/ecoli_info/roc_semppr_ecoli.csv')
elif organism == 'scer':
    rocsemppr_df = pd.read_csv('../Data/scer_info/roc_semppr_scer.csv')
print(rocsemppr_df.shape)
rocsemppr_df.head()

rocsemppr_df = rocsemppr_df.drop(columns=rocsemppr_df.columns[2:])
rocsemppr_df.columns = ['locus_tag', 'roc_semppr_mean']
df = df.merge(rocsemppr_df, on='locus_tag', how='outer')
print(df.shape)
df.head()

# Add in the Codon Adaptation Index (`CAI`)

The `Codon Adaptation Index` is one of the most famous codon usage bias metrics, originally proposed here:
https://academic.oup.com/nar/article-abstract/15/3/1281/1166844

It has a crazy number of citations and generally performs well at the task of predicting gene expression from sequence information. However, it requires some arbitrary and extraneous information. At its heart, the metric is a **distance** metric. It measures the distance of a given coding sequence (in terms of its codon usage bias) to a **pre-defined reference set of coding sequences** that are presumed to be highly expressed, and from which codon-level weights are calculated. Some work has been done to try to choose that reference set in an automated fashion from only knowledge in the genome, but to my knowledge nothing seems to outperform the pre-defined reference sets proposed in the paper (used below). And of course, this is to be expected since that reference set is a bit of a cheat since it already requires some outside gene expression knowledge and folds this knowledge into predictions of gene expression.

So I wrote some code to calculate weights starting from the reference set gene names (see below, commented). In the end, to ensure that I was getting things right or mostly right for *S. cerevisiae* I skipped this step and just copy/pasted the table of codon weights from the original paper into a `.txt` (tab-separated) file. 

###If you want to compute it yourself, but I'm just going to load the dictionaries from the orignal
###paper to make things consistent
# def get_codon_dicts(n=11):
#     """
#     Gets the codon to amino acid look up dictionary and the amino acid to 
#     (list of) codons dictionary using Biopython's CodonTable module
#     """
#     codon_to_aa = CodonTable.unambiguous_dna_by_id[n].forward_table
#     aa_to_codons = {}
#     for codon, aa in codon_to_aa.items():
#         try:
#             aa_to_codons[aa].append(codon)
#         except:
#             aa_to_codons[aa] = [codon]
#     return codon_to_aa, aa_to_codons 
    
# def flatten_lol(lol):
#     """
#     Merges a list of lists into a single large list
#     """
#     return [inner for outer in lol for inner in outer]

# def tripletize(cds_string):
#     """
#     Split a coding sequence into triplets (codons)
#     """
#     assert len(cds_string) % 3 == 0
#     return [cds_string[i:i+3] for i in range(0, len(cds_string), 3)]

# def calculate_RSCUs(codon_count_dict):
#     """
#     Calculates the relative synonymous codon usage values for a *SINGLE* dictionary
#     containing codon counts (codon:count) for *ONE* amino acid
#     """
#     n_total = sum(codon_count_dict.values())
#     unique_codons = len(codon_count_dict.keys())
#     rscu_dict = {}  
#     if n_total != 0:
#         for codon, codon_count in codon_count_dict.items():
#             rscu_dict[codon] = codon_count / (float(n_total) / unique_codons)
#     else:
#         for codon, codon_count in codon_count_dict.items():
#             rscu_dict[codon] = 0 
#     return rscu_dict

# def calculate_codon_weights(list_of_cds_strings, zero_val=0.001):
#     """
#     Calulates the weight of each codon from a reference set of genes
#     """
#     ###Get housekeeping dictionaries
#     codon_to_aa, aa_to_codons = get_codon_dicts()
    
#     if len(list_of_cds_strings) > 1: #when given a list of strings, make it one long list of codons
#         all_codons = []
#         for cds in list_of_cds_strings:
#             all_codons.append(tripletize(cds))
#         all_codons = flatten_lol(all_codons)
#     else:
#         all_codons = tripletize(list_of_cds_strings)

#     all_codon_counts = Counter(all_codons) #create dictionary of codon counts in the set
    
#     ###Workhorse of the code
#     weights_dict = {}
#     for aa, codons in aa_to_codons.items(): #iterates through single amino acids
#         single_aa_codon_counts = {k: all_codon_counts.get(k, None) for k in codons}
#         single_rscu_dict = calculate_RSCUs(single_aa_codon_counts) #get RSCU dictionary for the aa
#         max_rscu = max(single_rscu_dict.values()) #weights are normalized by the max RSCU
#         for codon, rscu_val in single_rscu_dict.items():
#             if single_rscu_dict[codon] == 0:
#                 weights_dict[codon] = zero_val
#             else:
#                 weights_dict[codon] = rscu_val / max_rscu
#     return weights_dict

# def calculate_cai(single_cds_str, weights_dict):
#     """
#     Calculates the CAI from a gene (string) and a dictionary of codon weight values.
#     """
#     codon_list = tripletize(single_cds_str)
#     weights_list = []
#     for codon in codon_list:
#         weights_list.append(weights_dict.get(codon, None))
#     weights_list =  [i for i in weights_list if i] 
#     cai = stats.gmean(weights_list)
#     return cai

# reference_set_loci = pd.read_csv('../Data/ecoli_info/original_CAI_refset.txt', header=None, names=['locus_tag'])
# ref_set_seqs = list(df.merge(reference_set_loci, on='locus_tag')['cds_seq'])
# cai_weights_dict = calculate_codon_weights(ref_set_seqs)

# cai_vals = []
# for index in df[df['well_behaved']==True].index:
#     cds_seq = df.loc[index]['cds_seq']
#     cai_vals.append(calculate_cai(cds_seq, cai_weights_dict))
    
# df.at[df[df['well_behaved']==True].index, 'CAI'] = cai_vals

**Defining a few simple functions to help out**

def tripletize(cds_string):
    """
    Split a coding sequence into triplets (codons)
    """
    assert len(cds_string) % 3 == 0
    return [cds_string[i:i+3] for i in range(0, len(cds_string), 3)]

def calculate_cai(single_cds_str, weights_dict):
    """
    Calculates the CAI from a gene (string) and a dictionary of codon weight values.
    """
    codon_list = tripletize(single_cds_str)
    weights_list = []
    for codon in codon_list:
        weights_list.append(weights_dict.get(codon, None))
    weights_list =  [i for i in weights_list if i] 
    cai = stats.gmean(weights_list)
    return cai

cai_weights_df = pd.read_csv('../Data/original_CAI_data.txt', header=0, index_col=0, sep='\t')
cai_weights_df.drop(index=['UAA', 'UGA', 'UAG'], inplace=True)
cai_weights_df = cai_weights_df.astype(float)
if organism == 'ecoli':
    cai_weights_dict = cai_weights_df['ecoli_w'].to_dict()
elif organism == 'scer':
    cai_weights_dict = cai_weights_df['scer_w'].to_dict()
else:
    raise Exception('Incorrect organism identifier')

cai_vals = []
for index in df[df['well_behaved']==True].index:
    cds_seq = df.loc[index]['cds_seq'].replace('T', 'U')
    cai_vals.append(calculate_cai(cds_seq, cai_weights_dict))
    
df.at[df[df['well_behaved']==True].index, 'CAI'] = cai_vals

# Add in knowledge of the `tAI` 

The so-called `tRNA adaptation index` is a very orthogonal metric that uses knowledge of the tRNA genes contained within the genome of interest to calculate a coding sequence specific metric of adaptation to this hypothetical tRNA pool. To my knowledge it was originally proposed here:
https://academic.oup.com/nar/article/32/17/5036/1333956

And I am taking the original codon weight values for *E. coli* from here:
https://www.pnas.org/content/107/8/3645

There was a subsequent re-implementation / variation that I thought about taking from:
https://academic.oup.com/bioinformatics/article/33/4/589/2593585

But since that method is woefully inadequate in terms of its accessibility and usability (matlab?!), I'd have to take the actual codon-level weights for this variant from yet another publication that (as best as I can tell) simply re-implemented the same basic model with more transparency:
https://genominfo.org/journal/view.php?doi=10.5808/GI.2018.16.4.e28

This is all a long wind up to say that I should probably re-implement the weight calculations from scratch one day. 

if organism == 'ecoli':    
    weights_df = pd.read_csv('../Data/ecoli_info/NC_000913.3.original_tAI_weights.csv', header=None)
    weights_df = weights_df[weights_df[1].isnull()==False]
    weights_df.columns = ['codon', 'value']
    weights_dict = pd.Series(weights_df['value'].values,index=weights_df['codon']).to_dict()
elif organism =='scer':
    weights_df = pd.read_csv('../Data/scer_info/scer.original_tAI_weights.csv', header=None)
    weights_df = weights_df[weights_df[1].isnull()==False]
    weights_df.columns = ['codon', 'value']
    weights_dict = pd.Series(weights_df['value'].values,index=weights_df['codon']).to_dict()
else:
    raise Exception('Incorrect organism identifier')
print(weights_dict)

tai_vals = []
for index in df[df['well_behaved']==True].index:
    cds_seq = df.loc[index]['cds_seq']
    codons = tripletize(str(cds_seq))
    tempy = []
    for codon in codons:
        tempy.append(weights_dict.get(codon, None))
    tempy = [i for i in tempy if i]
    tai_vals.append(stats.gmean(tempy))

df.at[df[df['well_behaved']==True].index, 'tAI'] = tai_vals

temp_df = df[df['well_behaved']==True]
for method1 in ['roc_semppr_mean', 'CAI', 'tAI']:
    print(method1)
    for method2 in ['roc_semppr_mean', 'CAI', 'tAI']:
        print(method2, stats.spearmanr(temp_df[method1], temp_df[method2]))
    print()


# Write data to a file

if organism == 'ecoli':
    df.to_csv('../Data/ecoli_info/current_ecoli_master_table.tsv', sep='\t', index=False)
elif organism == 'scer':
    df.to_csv('../Data/scer_info/current_scer_master_table.tsv', sep='\t', index=False)
else:
    raise Exception('Incorrect organism identifier')

# Scratch

print(df.shape)
df.head()

df['well_behaved'].value_counts()
