==> Code/3-explore_expression_and_growth.Rmd <==

---
title: "Exploring gene expression and growth"
author: "Mackenzie Johnson"
date: "12/2/2021"
output: html_document
---

This document is the R version of `4-explore_expression_and_growth.ipynb`, created for easier extension of the analysis created by Adam.


```{r, message=FALSE}

library(tidyverse)
library(here)
library(glue)
library(viridis)
library(cowplot)
library(ggtext)
library(patchwork)
library(reshape2)

```


## Defining constants and save directories

```{r, message=FALSE}

data_key <- tibble(
  set = c(
    'ecoli_full', 
    'ecoli_sparse', 
    'ecoli_neutral',
    'ecoli_neutral_wt',
    'ecoli_non_at_rich',
    'ecoli_prot',
    'scer_rna', 
    'scer_prot'
  ),
  meta_file = c(
    here("Data", "processed_data", "processed_metadata_ecoli.tsv"),
    here("Data", "processed_data", "processed_metadata_ecoli_SPARSE.tsv"),
    here("Data", "processed_data", "processed_metadata_ecoli_NEUTRAL.tsv"),
    here("Data", "processed_data", "processed_metadata_ecoli_NEUTRAL_WT.tsv"),
    here("Data", "processed_data", "processed_metadata_ecoli.tsv"),
    here("Data", "processed_data", "processed_metadata_ecoli_PROTEIN.tsv"),
    here("Data", "processed_data", "processed_metadata_scer_RNA.tsv"),
    here("Data", "processed_data", "processed_metadata_scer_PROTEIN.tsv")
  ),
  expression_file = c(
    here("Data", "processed_data", "processed_expression_ecoli.tsv"),
    here("Data", "processed_data", "processed_expression_ecoli_SPARSE.tsv"),
    here("Data", "processed_data", "processed_expression_ecoli_NEUTRAL.tsv"),
    here("Data", "processed_data", "processed_expression_ecoli_NEUTRAL_WT.tsv"),
    here("Data", "processed_data", "processed_expression_ecoli.tsv"),
    here("Data", "processed_data", "processed_expression_ecoli_PROTEIN.tsv"),
    here("Data", "processed_data", "processed_expression_scer_RNA.tsv"),
    here("Data", "processed_data", "processed_expression_scer_PROTEIN.tsv")
  ),
  summary_file = c(
    here("Data", "processed_data", "processed_expression_summary_ecoli.tsv"),
    here("Data", "processed_data", "processed_expression_summary_ecoli_SPARSE.tsv"),
    here("Data", "processed_data", "processed_expression_summary_ecoli_NEUTRAL.tsv"),
    here("Data", "processed_data", "processed_expression_summary_ecoli_NEUTRAL_WT.tsv"),
    here("Data", "processed_data", "processed_expression_summary_ecoli.tsv"),
    here("Data", "processed_data", "processed_expression_summary_ecoli_PROTEIN.tsv"),
    here("Data", "processed_data", "processed_expression_summary_scer_RNA.tsv"),
    here("Data", "processed_data", "processed_expression_summary_scer_PROTEIN.tsv")
  ),
  cub_file = c(
    here("Data", "ecoli_info", "current_ecoli_master_table.tsv"),
    here("Data", "ecoli_info", "current_ecoli_master_table.tsv"),
    here("Data", "ecoli_info", "current_ecoli_master_table.tsv"),
    here("Data", "ecoli_info", "current_ecoli_master_table.tsv"),
    here("Data", "ecoli_info", "current_ecoli_master_table.tsv"),
    here("Data", "ecoli_info", "current_ecoli_master_table.tsv"),
    here("Data", "scer_info", "current_scer_master_table.tsv"),
    here("Data", "scer_info", "current_scer_master_table.tsv")
  )
)

# select dataset for analysis
data_tag <- "ecoli_full" #"ecoli_full" 

```


## Read in the data

```{r, message=FALSE}

# get relevant file names for data set of interest
data_key %>% 
  filter(set == data_tag) -> analysis_files

# read in data
meta_df <- read_tsv(analysis_files$meta_file)

exp_df <- read_tsv(analysis_files$expression_file)

exp_sum_df <- read_tsv(analysis_files$summary_file) # keeping this in but this file is never used - stats calculated in R Markdown files now

full_df <- read_tsv(analysis_files$cub_file) %>% 
  select(locus_tag, gene)

```

```{r}

if (data_tag == "ecoli_non_at_rich") {
  
  # read in list of AT-rich genes from dos Reis et al. 2003
  at_rich_df <- read_csv(
    here('Data', 'ecoli_info', 'at_rich_genes.csv'),
    col_names = c('locus_tag', 'gene'),
    show_col_types = FALSE
  )
  
  full_df <- full_df %>% 
    mutate(at_rich = 
             case_when(
               locus_tag %in% at_rich_df$gene ~ TRUE,
               gene %in% at_rich_df$locus_tag ~ TRUE,
               TRUE ~ FALSE
             ))
  
  non_at_genes <- full_df %>% filter(at_rich == FALSE) %>% pull(locus_tag)
  
  # filter out AT-rich genes
  exp_df <- exp_df %>% 
    filter(`log-TPM` %in% non_at_genes)
  
  exp_sum_df <- exp_sum_df %>% 
    filter(`log-TPM` %in% non_at_genes)
  
}

```


## Explore the relationship between conditions

```{r, message=FALSE}

# get pearson's correlation coefficient for gene expression between all conditions
if (data_tag %in% c("ecoli_full", "ecoli_sparse", "ecoli_neutral", "ecoli_neutral_wt", "ecoli_non_at_rich")) {
  cond_cor_mat <- exp_df %>% 
    select(-`log-TPM`) %>% 
    cor(., method = "pearson")
} else {
  cond_cor_mat <- exp_df %>% 
    select(-`Gene`) %>% 
    cor(., method = "pearson")
}

dim(cond_cor_mat)
cond_cor_mat[1:4, 1:4]

```


```{r, message=FALSE}

# make heatmap for correlations across conditions
fig_exp_cond_corr <- melt(cond_cor_mat^2) %>% 
  ggplot(aes(x = Var1, y = Var2, fill = value)) +
  geom_tile() +
  coord_fixed() +
  scale_fill_viridis_c(
    option = "B", name = "*R*<sup>2</sup>"
  ) +
  theme_void() +
  theme(legend.title = element_markdown())

```


#### Example scatterplots of two highly correlated conditions

```{r, message=FALSE}

# find max pearson's r value between two conditions
test_mat <- cond_cor_mat
diag(test_mat) <- 0 # exclude self correlations
inds <- which(test_mat == max(test_mat), arr.ind = TRUE)

# store data, condition names, and replace column names for use across datasets
max_cond_corr_df <- tibble(
  exp_df[, inds[1]+1],
  exp_df[, inds[2]+1]
)
max_names <- colnames(max_cond_corr_df)
colnames(max_cond_corr_df) <- c("cond1", "cond2")

# store position for R^2 values in plot
if (data_tag %in% c("ecoli_full", "ecoli_sparse", "ecoli_neutral", "ecoli_neutral_wt", "ecoli_non_at_rich")) {
  x_pos = 3
  y_pos = 13.5
} else if (data_tag == "ecoli_prot") {
  x_pos = -2
  y_pos = 3.5 
} else if (data_tag == "scer_prot") {
  x_pos <- -4
  y_pos <- 7 
} else {
  x_pos <- 3.5
  y_pos <- 9
}

# create dataframe with R^2 values to plot
r2_max_text <- data.frame(
  label = glue(
    '*R*<sup>2</sup> = {round(cond_cor_mat[inds[1], inds[2]]*
    cond_cor_mat[inds[1], inds[2]], 3)}'),
  x = x_pos, 
  y = y_pos
)

# store axes limits
if (data_tag %in% c("ecoli_full", "ecoli_sparse", "ecoli_neutral", "ecoli_neutral_wt", "ecoli_non_at_rich")) {
  xy_min <- 0
  xy_max <- 15
} else if (data_tag %in% c("ecoli_prot", "scer_prot")) {
  xy_min <- NULL
  xy_max <- NULL
} else {
  xy_min <- 2.5
  xy_max <- 10
}

# visualize two conditions with largest correlation
fig_max <- max_cond_corr_df %>% 
  ggplot(aes(x = cond1, y = cond2)) +
  geom_point(color = "black", alpha = 0.3) +
  coord_fixed() +
  scale_x_continuous(
    limits = c(xy_min, xy_max), 
    expand = c(0.01, 0.01),
    name = glue('Condition {inds[1]} expression')
    # name = glue('Condition {inds[1]}: {max_names[1]}')
  ) +
  scale_y_continuous(
    limits = c(xy_min, xy_max),
    expand = c(0.02, 0.02),
    name = glue('Condition {inds[2]} expression')
  ) +
  geom_richtext(
    data = r2_max_text,
    aes(x = x, y = y,label = label),
    label.color = NA,
    inherit.aes = FALSE
  ) +
  theme_half_open() +
  background_grid()

max_names
# ecoli_full = "ytf__delydcI_ph5" "ytf__wt_ph5" 
# ecoli_sparse = "ytf__delyiaJ" "ytf__delyeiE"
# ecoli_neutral = "ytf__wt_ph5" "ytf__delydcI_ph5"
# ecoli_prot = "Chemostat u=0.35" "Chemostat u=0.20"
# scer_rna = "Gln_0.1_6.87_0.55" "NH4_0.1_7.5_0.5" 
# scer_prot = "Ile_0.1_7.5_9.81" "NH4_0.05_7.5_0.5"

```


#### Example scatterplots of two minimally correlated conditions

```{r, message=FALSE}

# find min pearson's r value between two conditions
inds_min <- which(cond_cor_mat == min(cond_cor_mat), arr.ind = TRUE)

# store data, condition names, and replace column names for use across datasets
min_cond_corr_df <- tibble(
  exp_df[, inds_min[1]+1],
  exp_df[, inds_min[2]+1]
)
min_names <- colnames(min_cond_corr_df)
colnames(min_cond_corr_df) <- c("cond1", "cond2") # rename for easier reproducibility

# store R^2 value for use in plot
r2_min_text <- data.frame(
  label = glue(
    '*R*<sup>2</sup> = {round(cond_cor_mat[inds_min[1], inds_min[2]]*
    cond_cor_mat[inds_min[1], inds_min[2]], 3)}'),
  x = x_pos, 
  y = y_pos
)

# visualize the conditions with the smallest correlation
fig_min <- min_cond_corr_df %>% 
  ggplot(aes(x = cond1, y = cond2)) +
  geom_point(color = "black", alpha = 0.3) +
  coord_fixed() +
  scale_x_continuous(
    limits = c(xy_min, xy_max),
    expand = c(0.01, 0.01),
    name = glue('Condition {inds_min[1]} expression')
  ) +
  scale_y_continuous(
    limits = c(xy_min, xy_max),
    expand = c(0.02, 0.02),
    name = glue('Condition {inds_min[2]} expression')
  ) +
  geom_richtext(
    data = r2_min_text,
    aes(x = x, y = y,label = label),
    label.color = NA,
    inherit.aes = FALSE
  ) +
  theme_half_open() +
  background_grid()

#fig_min

min_names
# ecoli_full = "ica__cytd_rib" "ssw__glc_xyl_glc"
# ecoli_spare = "ica__cytd_rib" "ssw__glc_xyl_glc"
# ecoli_neutral = "ssw__wt_xyl" "ica__cytd_rib"
# ecoli_prot = "pH6 glucose" "LB" 
# scer_rna = "NH4_0.35_7.5_0.5" "NH4_0.05_7.5_0.5"
# scer_prot = "Gln_0.1_7.5_5.53" "NH4_0.35_7.5_0.5"

```


### Figure 2 
```{r, message=FALSE}

fig_2 <- plot_grid(
  fig_exp_cond_corr, fig_min, fig_max,
  nrow = 1,
  labels = "AUTO",
  scale = c(1, 1, 1)
)

fig_2

```


```{r, include=FALSE, eval=FALSE}

# needs white background due to transparency of heatmap and outer rectangle
fig_2_fixed <- cowplot::ggdraw(fig_2) + 
  theme(plot.background = element_rect(fill="white", color = NA)) 

# save plot
save_plot(
  here(
    "Manuscript",
    "Figures",
    glue('fig_corr_{data_tag}.png')   # file name
  ),
  fig_2_fixed, 
  ncol = 1, 
  nrow = 1, 
  base_height = 3.71, #base_height = 5.71,
  base_asp = 2.618, # 1.618
  base_width = NULL
)

```


```{r, message=FALSE, include=FALSE}

# remove intermediate plots from memory
rm(fig_exp_cond_corr, fig_max, fig_min, fig_2b, fig_2)

# remove unnecessary data frames and values
rm(max_cond_corr_df, min_cond_corr_df, 
   r2_max_text, r2_min_text, test_mat,
   inds, inds_min, max_names, min_names,
   x_pos, y_pos, xy_min, xy_max)

```


## Explore growth rate data

```{r, message=FALSE}

meta_df %>% 
  ggplot(aes(x = `Growth Rate (1/hr)`)) +
  geom_histogram(binwidth = 0.05) +
  scale_x_continuous(
    limits = c(0, 1.5),
    expand = c(0,0)
  ) +
  scale_y_continuous(
    name = "Count", 
    expand = c(0,0)
  ) +
  theme_minimal_hgrid()

# slowest rate
slow_rate <- min(meta_df$`Growth Rate (1/hr)`)
# fastest rate
fast_rate <- max(meta_df$`Growth Rate (1/hr)`)
# slowest dt
slow_dt <- log(2)/slow_rate
# fastest dt
fast_dt <- log(2)/fast_rate

```

- Slowest rate: `r slow_rate`
- Fastest rate: `r fast_rate`
- Slowest doubling time: `r slow_dt`
- Fastest doubling time: `r fast_dt`

#### Assess individual gene correlations with the growth rate data

I'm only doing a single randomization/permutation here, and it should of course be noted that this randomization is a bit unfair since it destroys the underlying correlation structure between conditions. However, a permutation strategy that accounts for that structure would be very difficult and the only point here is to show that there are some genes that are more/less correlated with growth than you might expect by chance alone.

## Expression vs growth correlation

```{r, message=FALSE}

if (data_tag %in% c("ecoli_full", "ecoli_sparse", "ecoli_neutral", "ecoli_neutral_wt", "ecoli_non_at_rich")) {
  exp_gr_df <- exp_df %>% 
    gather(-`log-TPM`,
           key = "condition",
           value = "expression") %>% 
    spread(`log-TPM`, expression)
} else {
  exp_gr_df <- exp_df %>% 
    gather(-Gene,
           key = "condition",
           value = "expression") %>% 
    spread(Gene, expression)
}

exp_gr_df <- meta_df %>% 
    select(`Simple_sample_id`, `Growth Rate (1/hr)`) %>% 
    inner_join(., exp_gr_df, by = c("Simple_sample_id" = "condition"))

exp_gr_corr <- cor(exp_gr_df[2], exp_gr_df[3:ncol(exp_gr_df)], method = "pearson") %>% 
   as_tibble() %>% 
   gather(
     .,
     key = "genes",
     value = "pearsons"
    ) %>%
   mutate(r2 = pearsons^2)
  
perm_mat <- lapply(exp_gr_df[3:ncol(exp_gr_df)], sample) %>% as_tibble()
exp_gr_perm <- cor(exp_gr_df[2], perm_mat, method = "pearson") %>% 
  as_tibble() %>% 
  gather(
    .,
    key = "genes",
    value = "pearsons"
  ) %>% 
  mutate(r2 = pearsons^2)
  
perm_df <- tibble(
  exp_gr_corr$pearsons,
  exp_gr_perm$pearsons
) %>%
  gather(., "set", "pearsons")

exp_gr_corr %>% filter(pearsons == max(pearsons)) %>% .$genes
exp_gr_corr %>% filter(pearsons == min(pearsons)) %>% .$genes


```


```{r, message=FALSE}

make_cor_plot <- function(x, gname) {
  x <- enquo(x)
  exp_gr_df %>%
    ggplot() +
    aes(!!x, `Growth Rate (1/hr)`) +
    geom_point(alpha = 0.75) +
    scale_x_continuous("Expression across conditions") +
    theme_minimal_grid() +
    ggtitle(glue("{gname}"))
}

if (data_tag %in% c("ecoli_full", "ecoli_non_at_rich")) {
  
  fig_ant_cor <- make_cor_plot(b4658_2, "yghX")
  fig_pos_cor <- make_cor_plot(b2780, "pyrG")
  
} else if (data_tag == "ecoli_sparse") {
  
  fig_ant_cor <- make_cor_plot(b3586, "yiaV")
  fig_pos_cor <- make_cor_plot(b0170, "tsf")
  
} else if (data_tag == "ecoli_neutral") {
  
  fig_ant_cor <- make_cor_plot(b2106, "rcnA")
  fig_pos_cor <- make_cor_plot(b0154, "hemL")
  
} else if (data_tag == "ecoli_neutral_wt") {
  
  fig_ant_cor <- make_cor_plot(b1042, "csgA")
  fig_pos_cor <- make_cor_plot(b3789, "rffH")
  
} else if (data_tag == "ecoli_prot") {
  
  fig_ant_cor <- make_cor_plot(livJ, "livJ")
  fig_pos_cor <- make_cor_plot(kbl, "kbl")
  
} else if (data_tag == "scer_rna") {
  
  fig_ant_cor <- make_cor_plot(INM2, "INM2")
  fig_pos_cor <- make_cor_plot(GLN4, "GLN4")
  
} else {
  
  fig_ant_cor <- make_cor_plot(URA10, "URA10")
  fig_pos_cor <- make_cor_plot(RPS9A, "RPS9A")
  
}

fig_perm <- perm_df %>% 
  ggplot() +
    aes(x = pearsons, fill = set) +
    geom_density(alpha = 0.5) +
    scale_x_continuous(
      #name = "Pearson's *r* (individual gene expression vs growth)<br />**GCI**"
      name = "GCI: Pearson's *r* (individual gene expression vs growth)"
    ) +
    scale_y_continuous(name = "Density") +
    scale_fill_manual(
      name = "",
      labels = c("True data", "Shuffled data"),
      values = c("#009E73", "#808080")
    ) +
    theme_minimal_grid() +
    theme(
      legend.position = "top",
      axis.title.x = element_markdown()
    )

#fig_perm

t_test_results <- t.test(exp_gr_corr$pearsons, exp_gr_perm$pearsons)
t_test_results$p.value

```


## Figure 4:
```{r}

fig_4 <- (fig_ant_cor | fig_pos_cor) / fig_perm +
   plot_annotation(
     tag_levels = "A"
   )

fig_4

```

```{r, include=FALSE, eval=FALSE}

save_plot(
  here(
    "Manuscript",  
    "Figures",
    glue('fig_gene_con_{data_tag}.png')   # file name
  ),
  fig_4, 
  ncol = 1, 
  nrow = 1, 
  base_height = 6.71, #base_height = 5.71,
  base_asp = 1.618, 
  base_width = NULL
)

```


#### Save gene GCI data for use in GO analysis
```{r, message=FALSE, eval=FALSE}

if (data_tag %in% c("ecoli_full", "ecoli_sparse", "ecoli_neutral", "ecoli_neutral_wt", "ecoli_non_at_rich")) {
  
  genes_positive <- exp_gr_corr %>% 
    filter(pearsons > 0) %>% 
    select(genes, pearsons) %>% 
    left_join(., full_df, by = c("genes" = "locus_tag"))
  colnames(genes_positive) <- c("locus_tag", "GCI", "gene_name")

  genes_negative <- exp_gr_corr %>% 
    filter(pearsons < 0) %>% 
    select(genes, pearsons) %>% 
    left_join(., full_df, by = c("genes" = "locus_tag"))
  colnames(genes_negative) <- c("locus_tag", "GCI", "gene_name")
  
} else {
  
  genes_positive <- exp_gr_corr %>% 
    filter(pearsons > 0) %>% 
    select(pearsons, genes) %>% 
    filter(genes != "2020-10-01 00:00:00")
  colnames(genes_positive) <- c("GCI", "gene_name")
  
  genes_negative <- exp_gr_corr %>% 
    filter(pearsons < 0) %>% 
    select(pearsons, genes) %>% 
    filter(genes != "2020-10-01 00:00:00")
  colnames(genes_negative) <- c("GCI", "gene_name")
  
}

write_csv(
  genes_positive, 
  here(
    'Data',
    'genes_GCI',
    glue('{data_tag}_pos_corr.csv')
  ),
  append = FALSE
)

write_csv(
  genes_negative, 
  here(
    'Data',
    'genes_GCI',
    glue('{data_tag}_neg_corr.csv')
  ),
  append = FALSE
)

```


### Session Info
```{r}
sessionInfo()
```
